Create static library for libndt7
* Split `libndt7.hpp` into `libndt7.h` and `libndt7.cc`.
* Build the static library `libndt7.a`.
* Update the cmake file to install the static library and header.
